custom JSONEncoder with optional indent
